Fixed failure on user editing if email changed
The target user was being found by the email
the admin sent on the editing form. If the admin
decided to change the email the user couldn't be found,
or worse: another user would be edited.
Now, when the user is selected for editing, its id
is stored on a session variable wich is then used
to retrieve the original user to edit.